backend/requirements.txt
python>=3.9
pip>=21.0
python-dotenv
psycopg2-binary